Sync date() with En & translate DateTimeInterface::format() to save work
This not only solves next two errors from the list but also completes
one of the most scary items that was pending for years in the Polish
translation. Surprisingly it wasn't that hard after all and I
successfully moved 90% of the translations that used to reside in the
date() function's page.