Sync control-structures.xml with En, just one more file to go!
Now it's just a matter of re-syncing functions.xml and the Polish
translation of the PHP Manual should be building again.

A ton of work that I'm not sure if anyone needed but now whoever wants
to pick up will be soon greeted with either working or almost working
build.

At this point I cannot even estimate how many files that were previously
outdated or were missing the En-Revision value have been brought up to
date. It's definitely the biggest effort for the Polish PHP manual done
in at least last five years.