Translate some array-sorting entities using old translations
We are about to remove this information from the sort() function's page
but before doing so, let's reuse the useful bits in order to translate
the globally available entity describing the flags param for array
sorting functions.

Then we also translate two more entities used on various pages dealing
with array sorting.