Straight up copy English updates for some WIP entities
I'm not particularly happy about it but there are a few sections of the
language-enntities.ent that are left untranslated in the Polish tree,
mostly related to MongoDB.

This allows us to stay "in sync" with English, at least in revcheck's
terms. I very much prefer having some copy-pasted untraslated stuff
than multi-thousand diff on doc.php.net the next time someone tries
to update this hummoungous file.